Add missing break in a case statement.
When decoding IBS_OP_DATA2 packet types, cores from Family 10h
and from the "Bulldozer" cores (pre-Piledriver) could have a
request-source field that said the data came from a local L3.
This became a reserved encoding on later processors, so we
have not tested it very rigorously.

Unfortunately, we missed a break statement in the decoder for
this, meaning the CSV output file could have shifted fields
that also included the next possible decoding. This patch removes
this error.

Thanks to Stefano Carnà from Sapienza for this bug report.