Print machine setup information for fetch files too
Previously, we only printed out machine-specific information,
such as the OS version, timestamp of when the benchmark was run,
and working directory, in the op sample file. However, this should
be included in any sample file to help identify where they came
from if they are stored off for later use.