Add support for newer AMD processors, fix "Zen" workaround
"Zen" CPU cores are used in more processors than just the
CPU-only SoCs made of "Zeppelin" chips (Family 17h Models 00h-0Fh).
However, these processors that use the "Zen" CPU may still have IBS
disabled by default and require the same workarounds as models
00h-0Fh. This patch enables workarounds for these other processor
models.

In addition, this patch makes modifications to how the "Zen"
workarounds are done. Previously, we saw intermittent lockups when
doing "workaround 2". I originally thought this was because of
other work going on in the pipeline, so I added in a 'wbinvd' to
do a pipeline flush. This still caused locksup on driver installation
every so often. Further investigation shows that the problem is
likely because of work from the "other" thread in a core. As such,
this patch modifies how "workaround 2" is done. If two threads on
the same core are enabled when trying to do the workaround on a
core, the other hardware thread is disabled by hot-unplugging it.
After the workaround is applied to the core, the sibling is
brought back up. This appears to be a much more robust way to
install the driver on "Zen"-based SoCs.