Update how IBS op sampling is stopped to prevent unhandled NMIs.
The interrupt associated with IBS op samples is not instantaneous.
As such, there is an amount of time between when the processor
takes an IBS op sample and the time when the NMI handler starts
to work.

Because of this, it is possible that the driver will attempt to
disable IBS op sampling by writing IBS_OP_CTL such that IbsOpEn
is disabled. Unfortunately, the write to IBS_OP_CTL will overwrite
the value of IbsOpVal. Previously, we always set IbsOpVal to zero
whenever we disabled IBS ops ampling.

This meant that when the "delayed" NMI handler eventually started,
it would see IbsOpVal=0 and believe that the interrupt did not come
from the IBS hardware. It would pass the interrupt down the NMI
chain, which has the possibility of causing reboots (e.g. when the
NMI watchdog is enabled).

Reading IBS_OP_CTL to obtain the "old" value of IbsOpVal cannot
solve this problem, because there is a data race between the RDMSR
and WRMSR. You could read IbsOpVal as zero, and before the
subsequent write to IBS_OP_CTL, an IBS op sample would be taken.
The hardware would set IbsOpVal, then the subsequent write unsets
it, leading to the same problem described above.

As such, our solution is to always set IbsOpVal to 1 when disabling
op sampling. If there is a "delayed" NMI coming it, it will see the
sample as valid. Because IbsOpMaxCnt is set to 0, however, it will
believe this is a bad "delayed" sample and throw it away.
We need to throw this away in case the delayed NMI didn't actually
come from an IBS sample, but someone else on the NMI chain.
Otherwise, we could take a bad sample and save it out.

The down-side of this solution is precisely mentioned above -- we
could "handle" an NMI request that comes from someone else on the
NMI chain. To prevent our solution from "eating" the next NMI
no matter how far away it is, we have our IBS disable function wait
for a microsecond after turning off IBS and setting the potentially-
fake IbsOpVal (hopefully long enough for the delayed NMI to arrive)
and then forcing IbsOpVal back to 0 since hopefully no NMIs are
coming in. This reduces the size of the window where we incorrectly
eat someone else's interrupt.