Modify run_and_annotate tool to allow setting IBS sample rate
Previously, the IBS monitoring application could change the op and
fetch sampling rates, but the wrapper application
'ibs_run_and_annotate' would only use its default sampling rate.
This meant that users who wanted to change the sample rate would
need to continually modify either the script or the default in
the monitoring application (requiring a recompilation).

These changes allow an optional argument to the flags that turn
on op and fetch sampling in the run_and_annotate script. Now, we
default to the old rates but allow the user to pass in new sampling
rates if they want.